auth refactoring
