+++ src/hooks/useAuth.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { AuthContext } from "../contexts/AuthContext";
+
+export function useAuth(){
+    return useContext(AuthContext)
+}
