+++ src/pages/NewRoom.tsx
@@ -3,8 +3,11 @@ import logoImg from '../assets/images/logo.svg'
 import '../styles/auth.scss'
 import { Button } from '../components/Button'
 import {Link} from 'react-router-dom'
+//import { useAuth } from '../hooks/useAuth'
 
 export function NewRoom(){
+
+    //const {user} = useAuth()
 
     return (
         <div id="pageAuth">
